Better code management. Updated the notes in README.md
