+++ full_stack_python/full_stack_python.py
@@ -3,6 +3,7 @@
 import reflex as rx
 
 from rxconfig import config
+from .ui.base import base_page
 
 
 class State(rx.State):
@@ -11,88 +12,6 @@ class State(rx.State):
 
     def handle_heading_input_change(self, val):
         self.heading_addition = val
-
-
-def navbar_link(text: str, url: str) -> rx.Component:
-    return rx.link(
-        rx.text(text, size="4", weight="medium"), href=url
-    )
-
-
-def navbar() -> rx.Component:
-    return rx.box(
-        rx.desktop_only(
-            rx.hstack(
-                rx.hstack(
-                    rx.image(
-                        src="/logo.jpg",
-                        width="2.25em",
-                        height="auto",
-                        border_radius="25%",
-                    ),
-                    rx.heading(
-                        "Reflex", size="7", weight="bold",
-                    ),
-                    align_items="center",
-                ),
-                rx.hstack(
-                    navbar_link("Home", "/#"),
-                    navbar_link("About", "/#"),
-                    navbar_link("Pricing", "/#"),
-                    navbar_link("Contact", "/#"),
-                    justify="end",
-                    spacing="5",
-                ),
-                justify="between",
-                align_items="center",
-            ),
-        ),
-        rx.mobile_and_tablet(
-            rx.hstack(
-                rx.hstack(
-                    rx.image(
-                        src="/logo.jpg",
-                        width="2em",
-                        height="auto",
-                        border_radius="25%",
-                    ),
-                    rx.heading(
-                        "Reflex", size="6", weight="bold",
-                    ),
-                    align_items="center",
-                ),
-                rx.menu.root(
-                    rx.menu.trigger(
-                        rx.icon("menu", size=30),
-                    ),
-                    rx.menu.content(
-                        rx.menu.item("Home"),
-                        rx.menu.item("About"),
-                        rx.menu.item("Pricing"),
-                        rx.menu.item("Contact"),
-                    ),
-                    justify="end",
-                ),
-                justify="between",
-                align_items="center",
-            ),
-        ),
-        bg=rx.color("accent", 3),
-        padding="1em",
-        # position="fixed",
-        # top="0px",
-        # z_index="5",
-        width="100%",
-    )
-
-
-def base_page(child: rx.Component, hide_navbar: bool = False, *args, **kwargs) -> rx.Component:
-    return rx.container(
-        *([navbar()] if not hide_navbar else []), # Conditionally show navbar
-        rx.color_mode.button(position="top-right"),
-        child, # This is the dynamic content (i.e., the content that changes from page to page)
-        rx.logo(),
-    )
 
 
 def index() -> rx.Component:
