+++ full_stack_python/ui/nav.py
@@ -1,0 +1,78 @@
+import reflex as rx
+
+
+def navbar_link(text: str, url: str) -> rx.Component:
+    return rx.link(
+        rx.text(text, size="4", weight="medium"), href=url
+    )
+
+
+def navbar() -> rx.Component:
+    return rx.box(
+        rx.desktop_only(
+            rx.hstack(
+                rx.hstack(
+                    rx.image(
+                        src="/logo.jpg",
+                        width="2.25em",
+                        height="auto",
+                        border_radius="25%",
+                    ),
+                    rx.heading(
+                        "Reflex", size="7", weight="bold",
+                    ),
+                    align_items="center",
+                ),
+                rx.hstack(
+                    navbar_link("Home", "/#"),
+                    navbar_link("About", "/#"),
+                    navbar_link("Pricing", "/#"),
+                    navbar_link("Contact", "/#"),
+                    spacing="5",
+                ),
+                rx.hstack(
+                    rx.color_mode.button(),
+                    justify="end",
+                ),
+                justify="between",
+                align_items="center",
+            ),
+        ),
+        rx.mobile_and_tablet(
+            rx.hstack(
+                rx.hstack(
+                    rx.image(
+                        src="/logo.jpg",
+                        width="2em",
+                        height="auto",
+                        border_radius="25%",
+                    ),
+                    rx.heading(
+                        "Reflex", size="6", weight="bold",
+                    ),
+                    align_items="center",
+                ),
+                rx.menu.root(
+                    rx.menu.trigger(
+                        rx.icon("menu", size=30),
+                    ),
+                    rx.menu.content(
+                        rx.menu.item("Home"),
+                        rx.menu.item("About"),
+                        rx.menu.item("Pricing"),
+                        rx.menu.item("Contact"),
+                        rx.menu.item(rx.color_mode.button()),
+                    ),
+                    justify="end",
+                ),
+                justify="between",
+                align_items="center",
+            ),
+        ),
+        bg=rx.color("accent", 3),
+        padding="1em",
+        # position="fixed",
+        # top="0px",
+        # z_index="5",
+        width="100%",
+    )
